src/annotation/alembic/env.py
import os

# from ctypes.wintypes import tagRECT
from logging.config import fileConfig

from sqlalchemy import engine_from_config
from sqlalchemy import pool

from alembic import context
# import sys

# TODO: migrate to a model-based approach
# sys.path.insert(0, "./alembic/models")
import annotations_app.models.base as base_models
target_metadata = base_models.db.metadata

# this is the Alembic Config object, which provides
# access to the values within the .ini file in use.
config = context.config

# Interpret the config file for Python logging.
# This line sets up loggers basically.
fileConfig(config.config_file_name)


if not os.environ.get("SQLALCHEMY_URL"):
    # Example: "mssql+pymssql://sa:Your_password123@compliance_mssql/master?charset=utf8"
    # or "postgresql://complianceuser:compliancepassword@compliance_postgres/compliancedb"
    raise Exception("Please configure SQLALCHEMY_URL")

config.set_main_option("sqlalchemy.url", os.environ.get("SQLALCHEMY_URL"))

# add your model's MetaData object here
# for 'autogenerate' support
# from myapp import mymodel
# target_metadata = mymodel.Base.metadata
# from utils import get_db
# from flask import Flask

# app = Flask(__name__)
# with app.app_context():
#    target_metadata = get_db().metadata
# target_metadata = None

# other values from the config, defined by the needs of env.py,
# can be acquired:
# my_important_option = config.get_main_option("my_important_option")
# ... etc.


def run_migrations_offline():
    """Run migrations in 'offline' mode.

    This configures the context with just a URL
    and not an Engine, though an Engine is acceptable
    here as well.  By skipping the Engine creation
    we don't even need a DBAPI to be available.

    Calls to context.execute() here emit the given string to the
    script output.

    """
    url = config.get_main_option("sqlalchemy.url")
    context.configure(
        url=url,
        target_metadata=target_metadata,
        literal_binds=True,
        dialect_opts={"paramstyle": "named"},
    )

    with context.begin_transaction():
        context.run_migrations()


def run_migrations_online():
    """Run migrations in 'online' mode.

    In this scenario we need to create an Engine
    and associate a connection with the context.

    """
    connectable = engine_from_config(
        config.get_section(config.config_ini_section),
        prefix="sqlalchemy.",
        poolclass=pool.NullPool,
    )

    with connectable.connect() as connection:
        context.configure(
            connection=connection, target_metadata=target_metadata
        )

        with context.begin_transaction():
            context.run_migrations()


if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()
